Move table_from_row_data to gammapy.utils.table
